Move decoder to common code.
